requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
python-multipart
argon2_cffi